Renamed asset handlers to asset stores and moved them into models.py
