--- temba/assets/__init__.py
@@ -1,13 +1,0 @@
-from __future__ import absolute_import, unicode_literals
-
-from enum import Enum
-from .handlers import ContactExportAssetHandler, ResultsExportAssetHandler, MessageExportAssetHandler
-
-
-class AssetType(Enum):
-    contact_export = (ContactExportAssetHandler,)
-    results_export = (ResultsExportAssetHandler,)
-    message_export = (MessageExportAssetHandler,)
-
-    def __init__(self, handler_class):
-        self.handler = handler_class(self)
