+++ temba/contacts/models.py
@@ -1495,11 +1495,11 @@ class ExportContactsTask(SmartModel):
         temp.flush()
 
         # save as file asset associated with this task
-        from temba.assets import AssetType
+        from temba.assets.models import AssetType
         from temba.assets.views import get_asset_url
 
-        handler = AssetType.contact_export.handler
-        handler.save(self.pk, File(temp), 'csv' if use_csv else 'xls')
+        store = AssetType.contact_export.store
+        store.save(self.pk, File(temp), 'csv' if use_csv else 'xls')
 
         subject = "Your contacts export is ready"
         template = 'contacts/email/contacts_export_download'
